Fix un-include of ts source file when build while watching.
Removed out "files" property in tsconfig.json as it's the root cause of
problem to not include .ts source file in building. Thus editing won't
apply changes.

This in turn fixed the problem of error generated in IDE about
experimental supports with decorators in pb-spinwheel.component.ts with
the inclusion of the following line

let Playbasis: any = _;

and whereby

import * as _ from 'playbasis.js';

to make it works.

--

Also added a better debug log function accepting multiple arguments with
prefix of "[pb-spinwheel]" message.