Port all code. Fix setting property via template code.
Object and array needed to be sent to [<propery-name>] in template code.
Otherwise, send in normal without []. This is meant to let it interpret
conveniently without us to do anything further, or define a local
variable in a class that set them.

Spinwheel's flow begins properly and response comes back as expected.

Still rewards images are not rendered yet on spinwheel. This will be
fixed next.